修复浮点型隐式转换成整型丢失小数导致ROOT_NODE与CANVAS_NODE尺寸不一致问题

+++ frameworks/core/pipeline/pipeline_base.h
@@ -311,7 +311,7 @@ public:
         return nullptr;
     }
 
-    void SetRootSize(double density, double width, double height);
+    void SetRootSize(double density, float width, float height);
 
     void UpdateFontWeightScale();
